+++ lance/CMakeLists.txt
@@ -67,6 +67,7 @@ ament_target_dependencies(robot_status
 set(robot_controller_srcs
     "src/control_node.cpp"
     "src/robot/tf_cache.cpp"
+    "src/robot/path_sampler.cpp"
     "src/robot/robot_params.cpp"
     "src/robot/collection_state.cpp"
     "src/robot/robot_controller.cpp"
